Commit for issues #4 and #8. Implement caching with cached crate. Add multi dimension array computation with IxDyn.
