+++ Cargo.toml
@@ -5,6 +5,7 @@ authors = ["clement"]
 edition = "2018"
 
 [dependencies]
+cached = "0.8"
 docopt = "1.0"
 ndarray = "0.12"
 ndarray-parallel = "0.9"
